[ui-components] refactor: moved colorpicker into ui package
