+++ packages/plugin-commons/web/package.json
@@ -19,14 +19,14 @@
     "libs",
     "viewer"
   ],
-  "main": "dist/cjs/index.js",
+  "main": "dist/module.cjs.js",
   "types": "dist/src/index.d.ts",
-  "module": "dist/es/index.js",
+  "module": "dist/module.js",
   "style": "dist/styles.min.scss",
   "scripts": {
     "clean": "rm -rf dist/",
     "prebuild": "npm run clean",
-    "build": "rollup --environment DYNAMIC_IMPORT,MODULE_NAME:plugin-commons -c ../../../rollup.config.js",
+    "build": "rollup --environment MODULE_NAME:plugin-commons -c ../../../rollup.config.js",
     "build:dev": "cross-env NODE_ENV=development npm run build",
     "build:analyze:editor": "npm run build -- --environment MODULE_ANALYZE_EDITOR  && open stats.html",
     "build:analyze:viewer": "npm run build -- --environment MODULE_ANALYZE_VIEWER && open stats.html",
@@ -34,7 +34,6 @@
     "lint": "eslint 'src/**/*.{js,jsx,ts,tsx,d.ts}'"
   },
   "dependencies": {
-    "react-colorful": "^4.4.3",
     "react-measure": "~2.3.0",
     "wix-rich-content-common": "8.47.7",
     "wix-rich-content-editor-common": "8.47.7",
