+++ packages/plugin-commons/web/src/Icons/index.ts
@@ -20,4 +20,3 @@ export { default as EditIcon } from './EditIcon';
 export { default as RTLIcon } from './RTLIcon';
 export { default as LTRIcon } from './LTRIcon';
 export { default as LoaderIcon } from './LoaderIcon';
-export { default as HashtagIcon } from './HashtagIcon';
