+++ packages/plugin-commons/web/src/index.ts
@@ -31,6 +31,5 @@ export {
 } from './Base/createBaseMediaPlugin';
 
 // Components
-export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
 
 export { getSelectionStyles } from 'wix-rich-content-editor-common';
