+++ packages/plugin-social-polls/web/src/components/settings/design-settings-section/DesignSettingsSection.jsx
@@ -1,8 +1,11 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-
-import { ColorPicker } from 'wix-rich-content-plugin-commons';
-import { SliderWithInput, Separator, SelectionList } from 'wix-rich-content-ui-components';
+import {
+  SliderWithInput,
+  Separator,
+  SelectionList,
+  ColorPicker,
+} from 'wix-rich-content-ui-components';
 import { mergeStyles } from 'wix-rich-content-common';
 
 import { ColorIcon, ImageIcon, GradientIcon } from '../../../assets/icons';
