+++ packages/plugin-text-color/web/package.json
@@ -38,7 +38,8 @@
     "react-modal": "~3.4.1",
     "wix-rich-content-common": "8.47.7",
     "wix-rich-content-editor-common": "8.47.7",
-    "wix-rich-content-plugin-commons": "8.47.7"
+    "wix-rich-content-plugin-commons": "8.47.7",
+    "wix-rich-content-ui-components": "8.47.7"
   },
   "unpkg": true,
   "publishConfig": {
