+++ packages/plugin-text-color/web/src/toolbar/TextColorPanel.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { ColorPicker } from 'wix-rich-content-plugin-commons';
+import { ColorPicker } from 'wix-rich-content-ui-components';
 import { Modifier, EditorState, getSelectionStyles } from 'wix-rich-content-editor-common';
 import { DEFAULT_STYLE_SELECTION_PREDICATE } from '../constants';
 import { getColor } from '../text-decorations-utils';
