+++ packages/toolbars/web/package.json
@@ -36,7 +36,8 @@
   "dependencies": {
     "wix-rich-content-common": "8.47.7",
     "wix-rich-content-editor-common": "8.47.7",
-    "wix-rich-content-plugin-commons": "8.47.7"
+    "wix-rich-content-plugin-commons": "8.47.7",
+    "wix-rich-content-ui-components": "8.47.7"
   },
   "unpkg": true,
   "publishConfig": {
