+++ packages/ui-components/web/src/Components/ColorPicker/CustomColorPicker.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { mergeStyles } from 'wix-rich-content-common';
-import Styles from '../../../statics/styles/custom-color-picker.scss';
+import Styles from '../../../statics/styles/color-picker-styles/custom-color-picker.scss';
 import { HashtagIcon } from '../../Icons';
 import { HexColorPicker, HexColorInput } from 'react-colorful';
 import classNames from 'classnames';
