+++ packages/ui-components/web/src/Components/ColorPicker/DefaultColorPicker.js
