+++ packages/ui-components/web/src/Icons/AddColorIcon.tsx
