+++ packages/ui-components/web/src/Icons/HashtagIcon.jsx
