+++ packages/ui-components/web/statics/styles/color-picker-styles/color-picker.scss
