+++ packages/ui-components/web/statics/styles/color-picker-styles/custom-color-picker-dialog.scss
